+++ block/initchain.go
@@ -25,7 +25,6 @@ func (m *Manager) RunInitChain(ctx context.Context) error {
 	// update the state with only the consensus pubkey
 	m.Executor.UpdateStateAfterInitChain(&m.State, res, gensisValSet)
 	m.Executor.UpdateMempoolAfterInitChain(&m.State)
-
 	if _, err := m.Store.UpdateState(m.State, nil); err != nil {
 		return err
 	}
