+++ block/types.go
@@ -2,6 +2,9 @@ package block
 
 import (
 	"github.com/dymensionxyz/dymint/types"
+
+	"github.com/libp2p/go-libp2p/core/crypto"
+	tmcrypto "github.com/tendermint/tendermint/crypto"
 )
 
 // TODO: move to types package
@@ -21,4 +24,12 @@ type blockMetaData struct {
 type CachedBlock struct {
 	Block  *types.Block
 	Commit *types.Commit
+}
+
+func getAddress(key crypto.PrivKey) ([]byte, error) {
+	rawKey, err := key.GetPublic().Raw()
+	if err != nil {
+		return nil, err
+	}
+	return tmcrypto.AddressHash(rawKey), nil
 }
